fix(tests): split branch validation tests to resolve LOC limit
Split 9 branch validation tests from test_task_service_fresh_db.py
into dedicated test_task_service_branch_validation.py file.

Original file: 501 lines → 418 lines
New file: 91 lines

All 17 tests pass (12 in original + 5 in new file), ruff ✓
